refactor button
