+++ frontend/src/Common/Buttons/ReloadButton.js
@@ -3,12 +3,12 @@ import PropTypes from 'prop-types';
 
 const reload_symbol = String.fromCharCode( 8635);
 const ReloadButton = (props) => (
-  <button type="button" onClick={props.setPicUrl} title={props.title} >
+  <button type="button" onClick={props.onClick} title={props.title} >
     {reload_symbol}
   </button>
 );
 ReloadButton.propTypes = {
   title: PropTypes.string.isRequired,
-  setPicUrl: PropTypes.func.isRequired
+  onClick: PropTypes.func.isRequired
 };
 export default ReloadButton;
