+++ frontend/src/Common/Buttons/index.js
@@ -1,4 +1,5 @@
 import CrossButton from './CrossButton';
 import PlusButton from './PlusButton';
+import ReloadButton from './ReloadButton';
 
-export {CrossButton, PlusButton};
+export {CrossButton, PlusButton, ReloadButton};
